CMS-49082 Refactor scripts and extends paths with auth
